base,prepare-sources: add source preparation hook
It's a common pattern among recipes to need to tweak the source tree after
fetch/unpack/patch but before configuration. This class provides a hook for
this use case. If recipes define do_prepare_sources, this task will be run
after unpack/patch but before configure. This avoids the need for the recipe
to add their own task, prepend configure, or use prefuncs/postfuncs to do so.

[YOCTO #6372]

Signed-off-by: Christopher Larson <chris_larson@mentor.com>