connman: also handle nfsroot for systemd
Install Andreas Oberritter's connmand-nfsroot wrapper script into
`${libexecdir}/connman` and adjust the startup script and systemd unit file to
run it. Sysvinit already used this logic, which ensures that connman doesn't
mess with a network interface which was used for nfsroot, so also apply that
logic for systemd, by using a wrapper script both can call.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>